refactor registry key to own file
