+++ src/main/java/org/cryptomator/windows/filemanagersidebar/ExplorerSidebarService.java
@@ -3,6 +3,7 @@ package org.cryptomator.windows.filemanagersidebar;
 import org.cryptomator.integrations.common.OperatingSystem;
 import org.cryptomator.integrations.common.Priority;
 import org.cryptomator.integrations.filemanagersidebar.SidebarService;
+import org.cryptomator.windows.common.RegistryKey;
 import org.cryptomator.windows.common.WindowsRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -28,8 +29,7 @@ public class ExplorerSidebarService implements SidebarService {
 		LOG.debug("Creating sidebar entry with CLSID {}", clsid);
 		//1. reg add HKCU\Software\Classes\CLSID\{0672A6D1-A6E0-40FE-AB16-F25BADC6D9E3} /ve /t REG_SZ /d "MyCloudStorageApp" /f
 		try (var t = WindowsRegistry.startTransaction()) {
-			try (var baseKey = t.createRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "Software\\Classes\\CLSID\\" + clsid, true)) {
-
+			try (var baseKey = t.createRegKey(RegistryKey.HKEY_CURRENT_USER, "Software\\Classes\\CLSID\\" + clsid, true)) {
 				baseKey.setStringValue("", entryName, false);
 
 				//2. reg add HKCU\Software\Classes\CLSID\{0672A6D1-A6E0-40FE-AB16-F25BADC6D9E3}\DefaultIcon /ve /t REG_EXPAND_SZ /d %%SystemRoot%%\system32\imageres.dll,-1043 /f
@@ -74,13 +74,13 @@ public class ExplorerSidebarService implements SidebarService {
 
 			//11. reg add HKCU\Software\Microsoft\Windows\CurrentVersion\Explorer\Desktop\NameSpace\{0672A6D1-A6E0-40FE-AB16-F25BADC6D9E3} /ve /t REG_SZ /d MyCloudStorageApp /f
 			var nameSpaceSubKey = "Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\Desktop\\NameSpace\\" + clsid;
-			try (var nameSpaceKey = t.createRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, nameSpaceSubKey, true)) {
+			try (var nameSpaceKey = t.createRegKey(RegistryKey.HKEY_CURRENT_USER, nameSpaceSubKey, true)) {
 				nameSpaceKey.setStringValue("", entryName, false);
 				LOG.trace("Created RegKey {} and setting default value", nameSpaceKey);
 			}
 
 			//12. reg add HKCU\Software\Microsoft\Windows\CurrentVersion\Explorer\HideDesktopIcons\NewStartPanel /v {0672A6D1-A6E0-40FE-AB16-F25BADC6D9E3} /t REG_DWORD /d 0x1 /f
-			try (var newStartPanelKey = t.createRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\HideDesktopIcons\\NewStartPanel", true)) {
+			try (var newStartPanelKey = t.createRegKey(RegistryKey.HKEY_CURRENT_USER, "Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\HideDesktopIcons\\NewStartPanel", true)) {
 				newStartPanelKey.setDwordValue(clsid, 0x1);
 				LOG.trace("Set value {} for RegKey {}", clsid, newStartPanelKey);
 			}
@@ -97,21 +97,21 @@ public class ExplorerSidebarService implements SidebarService {
 			try (var t = WindowsRegistry.startTransaction()) {
 				//undo step 11.
 				var nameSpaceSubkey = "Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\Desktop\\NameSpace\\" + clsid;
-				t.deleteRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, nameSpaceSubkey);
 				LOG.trace("Removing RegKey {}", nameSpaceSubkey);
+				t.deleteRegKey(RegistryKey.HKEY_CURRENT_USER, nameSpaceSubkey);
 
 				//undo step 12.
-				try (var nameSpaceKey = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\HideDesktopIcons\\NewStartPanel")) {
+				try (var nameSpaceKey = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "Software\\Microsoft\\Windows\\CurrentVersion\\Explorer\\HideDesktopIcons\\NewStartPanel")) {
 					LOG.trace("Removing Value {} of RegKey {}", clsid, nameSpaceKey);
 					nameSpaceKey.deleteValue(clsid);
 				}
 
 				//undo everything else
-				try (var baseKey = t.createRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "Software\\Classes\\CLSID\\{%s}".formatted(clsid), true)) {
+				try (var baseKey = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "Software\\Classes\\CLSID\\{%s}".formatted(clsid))) {
 					LOG.trace("Wiping everything under RegKey {} and key itself.", baseKey);
 					baseKey.deleteAllValuesAndSubtrees();
 				}
-				t.deleteRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "Software\\Classes\\CLSID\\{%s}".formatted(clsid));
+				t.deleteRegKey(RegistryKey.HKEY_CURRENT_USER, "Software\\Classes\\CLSID\\{%s}".formatted(clsid));
 				t.commit();
 			}
 		}
