+++ src/test/java/org/cryptomator/windows/common/WindowsRegistryIT.java
@@ -1,7 +1,6 @@
 package org.cryptomator.windows.common;
 
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
@@ -17,7 +16,7 @@ public class WindowsRegistryIT {
 	public void testOpenNotExisting() {
 		Assertions.assertThrows(RuntimeException.class, () -> {
 			try (var t = WindowsRegistry.startTransaction();
-				 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "i\\do\\not\\exist")) {
+				 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "i\\do\\not\\exist")) {
 
 			}
 		});
@@ -28,12 +27,12 @@ public class WindowsRegistryIT {
 	@Order(1)
 	public void testCreateNotExistingRollback() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.createRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win", true)) {
+			 var k = t.createRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win", true)) {
 		}
 
 		Assertions.assertThrows(RuntimeException.class, () -> {
 			try (var t = WindowsRegistry.startTransaction();
-				 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+				 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			}
 		});
 	}
@@ -43,12 +42,12 @@ public class WindowsRegistryIT {
 	@Order(2)
 	public void testCreateNotExistingCommit() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.createRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win", true)) {
+			 var k = t.createRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win", true)) {
 			t.commit();
 		}
 
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 		}
 	}
 
@@ -57,14 +56,14 @@ public class WindowsRegistryIT {
 	@Order(3)
 	public void testOpenSetValueRollback() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			k.setStringValue("exampleStringValue", "In Progress", false);
 		}
 
 		//TODO: be more specific in the assertion
 		Assertions.assertThrows(RuntimeException.class, () -> {
 			try (var t = WindowsRegistry.startTransaction();
-				 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+				 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 				k.getStringValue("exampleStringValue", false);
 			}
 		});
@@ -75,14 +74,14 @@ public class WindowsRegistryIT {
 	@Order(4)
 	public void testOpenSetValueCommit() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			k.setStringValue("exampleStringValue", "In Progress", false);
 			k.setDwordValue("exampleDwordValue", 0x42);
 			t.commit();
 		}
 
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			var stringData = k.getStringValue("exampleStringValue", false);
 			var binaryData = k.getDwordValue("exampleDwordValue");
 			Assertions.assertEquals("In Progress", stringData);
@@ -95,12 +94,12 @@ public class WindowsRegistryIT {
 	@Order(5)
 	public void testOpenDeleteValueRollback() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			k.deleteValue("exampleDwordValue");
 		}
 
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			var data = k.getDwordValue("exampleDwordValue");
 			Assertions.assertEquals(0x42, data);
 		}
@@ -111,13 +110,13 @@ public class WindowsRegistryIT {
 	@Order(6)
 	public void testOpenDeleteValueCommit() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			k.deleteValue("exampleDwordValue");
 			t.commit();
 		}
 
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			Assertions.assertThrows(RuntimeException.class, () -> {
 				k.getDwordValue("exampleDwordValue");
 			});
@@ -129,13 +128,13 @@ public class WindowsRegistryIT {
 	@Order(7)
 	public void testOpenDeleteTreeRollback() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			k.deleteAllValuesAndSubtrees();
 		}
 
 		Assertions.assertDoesNotThrow(() -> {
 			try (var t = WindowsRegistry.startTransaction();
-				 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+				 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 				k.getStringValue("exampleStringValue", false);
 			}
 		});
@@ -146,14 +145,14 @@ public class WindowsRegistryIT {
 	@Order(8)
 	public void testOpenDeleteTreeCommit() {
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win");
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win");
 			 var subk = t.createRegKey(k, "subkey", true)) {
 			k.deleteAllValuesAndSubtrees();
 			t.commit();
 		}
 
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 
 			Assertions.assertThrows(RuntimeException.class, () -> t.openRegKey(k, "subkey"));
 			Assertions.assertThrows(RuntimeException.class, () -> k.getStringValue("exampleStringValue", false));
@@ -165,11 +164,11 @@ public class WindowsRegistryIT {
 	@Order(9)
 	public void testDeleteRollback() {
 		try (var t = WindowsRegistry.startTransaction()) {
-			t.deleteRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win");
+			t.deleteRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win");
 		}
 
 		try (var t = WindowsRegistry.startTransaction();
-			 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+			 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 		}
 	}
 
@@ -178,13 +177,13 @@ public class WindowsRegistryIT {
 	@Order(10)
 	public void testDeleteCommit() {
 		try (var t = WindowsRegistry.startTransaction()) {
-			t.deleteRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win");
+			t.deleteRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win");
 			t.commit();
 		}
 
 		Assertions.assertThrows(RuntimeException.class, () -> {
 			try (var t = WindowsRegistry.startTransaction();
-				 var k = t.openRegKey(WindowsRegistry.RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
+				 var k = t.openRegKey(RegistryKey.HKEY_CURRENT_USER, "org.cryptomator.integrations-win")) {
 			}
 		});
 	}
